requirements.txt
gitPython
numpy
pandas